requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.72
botocore==1.35.72
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.3
django-cors-headers==4.3.1
django-filter==23.4
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
Events==0.5
google==3.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
inflection==0.5.1
jiter==0.8.0
jmespath==1.0.1
kombu==5.5.1
lxml==5.3.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.55.3
opensearch==0.9.2
opensearch-py==2.8.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.4
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
